backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
requests==2.26.0
pydantic==1.9.0
openai==0.11.3
Werkzeug==2.0.3
python-dotenv==1.0.0
